Made ThemeModeConverter public and refactored classes
